+++ src/universal/Router/Parameter.php
@@ -63,7 +63,7 @@ class Parameter implements ParameterInterface
      * @param string $name
      * @param bool   $hasDefaultValue
      * @param mixed  $defaultValue
-     * @param $classHinted
+     * @param \ReflectionClass $classHinted
      *
      * @throws \InvalidArgumentException when $classHinted is invalid (not a \ReflectionClass or null value
      */
@@ -118,6 +118,10 @@ class Parameter implements ParameterInterface
      */
     public function getClass(): \ReflectionClass
     {
+        if (!$this->classHinted instanceof \ReflectionClass) {
+            throw new \RuntimeException('Error this parameter '.$this->name.' has not class hinted');
+        }
+
         return $this->classHinted;
     }
 }